+++ src/android/api/org/signal/ringrtc/GroupCall.java
@@ -132,6 +132,7 @@ public final class GroupCall {
                             @Nullable ProxyInfo             proxyInfo,
                             @NonNull  byte[]                hkdfExtraInfo,
                             @Nullable Integer               audioLevelsIntervalMs,
+                            @Nullable Byte                  dredDuration,
                             @NonNull  PeerConnectionFactory factory,
                             @NonNull  Observer              observer) {
         Log.i(TAG, "create():");
@@ -139,6 +140,7 @@ public final class GroupCall {
         GroupCall call = new GroupCall(Kind.SIGNAL_GROUP, nativeCallManager, factory, observer);
 
         int audioLevelsIntervalMillis = audioLevelsIntervalMs == null ? 0 : audioLevelsIntervalMs.intValue();
+        byte dredDurationByte = dredDuration == null ? 0 : dredDuration.byteValue();
         try {
             call.clientId = ringrtcCreateGroupCallClient(
                 nativeCallManager,
@@ -147,6 +149,7 @@ public final class GroupCall {
                 proxyInfo,
                 hkdfExtraInfo,
                 audioLevelsIntervalMillis,
+                dredDurationByte,
                 // Returns a borrowed RC.
                 factory.getNativePeerConnectionFactory(),
                 // Returns a borrowed RC.
@@ -179,6 +182,7 @@ public final class GroupCall {
                             @Nullable byte[]                adminPasskey,
                             @NonNull  byte[]                hkdfExtraInfo,
                             @Nullable Integer               audioLevelsIntervalMs,
+                            @Nullable Byte                  dredDuration,
                             @NonNull  PeerConnectionFactory factory,
                             @NonNull  Observer              observer) {
         Log.i(TAG, "create():");
@@ -186,6 +190,7 @@ public final class GroupCall {
         GroupCall call = new GroupCall(Kind.CALL_LINK, nativeCallManager, factory, observer);
 
         int audioLevelsIntervalMillis = audioLevelsIntervalMs == null ? 0 : audioLevelsIntervalMs.intValue();
+        byte dredDurationByte = dredDuration == null ? 0 : dredDuration.byteValue();
         try {
             call.clientId = ringrtcCreateCallLinkCallClient(
                 nativeCallManager,
@@ -197,6 +202,7 @@ public final class GroupCall {
                 adminPasskey,
                 hkdfExtraInfo,
                 audioLevelsIntervalMillis,
+                dredDurationByte,
                 // Returns a borrowed RC.
                 factory.getNativePeerConnectionFactory(),
                 // Returns a borrowed RC.
@@ -1305,6 +1311,7 @@ public final class GroupCall {
                                           ProxyInfo proxyInfo,
                                           byte[] hkdfExtraInfo,
                                           int audioLevelsIntervalMillis,
+                                          byte dredDuration,
                                           long nativePeerConnectionFactory,
                                           long nativeAudioTrack,
                                           long nativeVideoTrack)
@@ -1320,6 +1327,7 @@ public final class GroupCall {
                                              byte[] adminPasskey,
                                              byte[] hkdfExtraInfo,
                                              int audioLevelsIntervalMillis,
+                                             byte dredDuration,
                                              long nativePeerConnectionFactory,
                                              long nativeAudioTrack,
                                              long nativeVideoTrack)