+++ src/rust/src/bin/call_sim-cli/endpoint/group_call_sim.rs
@@ -108,6 +108,7 @@ impl CallEndpoint {
                     None, // proxy_info
                     hkdf_extra_info,
                     Some(Duration::from_millis(200)),
+                    0,
                     Some(state.peer_connection_factory.clone()),
                     state.outgoing_audio_track.clone(),
                     state.outgoing_video_track.clone(),